+++ client/src/actions/toasts.js
@@ -1,5 +1,5 @@
-import { createError, createToast } from './factories';
-import { ADD_TOAST, REMOVE_TOAST, TOAST_FADE_IN, TOAST_FADE_OUT, ADD_ERROR, REMOVE_ERROR, ERROR_LOGGED } from './constants';
+import { createToast } from '../factories';
+import { ADD_TOAST, REMOVE_TOAST, TOAST_FADE_IN, TOAST_FADE_OUT } from '../constants';
 
 export function addToast(options = {}) {
     return (dispatch) => {
@@ -28,27 +28,5 @@ export function removeToast(id) {
     return {
         payload: id,
         type: REMOVE_TOAST
-    };
-}
-
-export function addError(options = {}) {
-    const error = createError(options);
-    return {
-        payload: error,
-        type: ADD_ERROR
-    };
-}
-
-export function removeError(id) {
-    return {
-        payload: id,
-        type: REMOVE_ERROR
-    };
-}
-
-export function errorLogged(id) {
-    return {
-        payload: id,
-        type: ERROR_LOGGED
     };
 }
