+++ client/src/constants.js
@@ -5,4 +5,3 @@ export const TOAST_FADE_OUT = 'TOAST_FADE_OUT';
 
 export const ADD_ERROR = 'ADD_ERROR';
 export const REMOVE_ERROR = 'REMOVE_ERROR';
-export const ERROR_LOGGED = 'ERROR_LOGGED';
